// src/lib/join-calendar/types.ts
export interface Persona {
	id: string;
	/**
	 * If true, the user's profile will be visible to the public.
	 */
	privacySettings?: boolean | null;
	name: string;
	email?: string | null;
	avatar?: (string | null) | Media;
	bio?: string | null;
	/**
	 * The skills of the user.
	 */
	skills?: (string | Topic)[] | null;
	/**
	 * Social links
	 */
	links?:
		| {
				url?: string | null;
				label?: string | null;
				id?: string | null;
		  }[]
		| null;
	user?: {
		docs?: (string | User)[];
		hasNextPage?: boolean;
		totalDocs?: number;
	};
	slug?: string | null;
	slugLock?: boolean | null;
	meta?: {
		title?: string | null;
		/**
		 * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
		 */
		image?: (string | null) | Media;
		description?: string | null;
	};
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
	id: string;
	alt: string;
	blurHash?: string | null;
	prefix?: string | null;
	updatedAt: string;
	createdAt: string;
	url?: string | null;
	thumbnailURL?: string | null;
	filename?: string | null;
	mimeType?: string | null;
	filesize?: number | null;
	width?: number | null;
	height?: number | null;
	focalX?: number | null;
	focalY?: number | null;
	sizes?: {
		'thumbnail-square'?: {
			url?: string | null;
			width?: number | null;
			height?: number | null;
			mimeType?: string | null;
			filesize?: number | null;
			filename?: string | null;
		};
		thumbnail?: {
			url?: string | null;
			width?: number | null;
			height?: number | null;
			mimeType?: string | null;
			filesize?: number | null;
			filename?: string | null;
		};
		medium?: {
			url?: string | null;
			width?: number | null;
			height?: number | null;
			mimeType?: string | null;
			filesize?: number | null;
			filename?: string | null;
		};
		large?: {
			url?: string | null;
			width?: number | null;
			height?: number | null;
			mimeType?: string | null;
			filesize?: number | null;
			filename?: string | null;
		};
		xlarge?: {
			url?: string | null;
			width?: number | null;
			height?: number | null;
			mimeType?: string | null;
			filesize?: number | null;
			filename?: string | null;
		};
		og?: {
			url?: string | null;
			width?: number | null;
			height?: number | null;
			mimeType?: string | null;
			filesize?: number | null;
			filename?: string | null;
		};
	};
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "topic".
 */
export interface Topic {
	id: string;
	name: string;
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
	id: string;
	/**
	 * The phone number of the user.
	 */
	phone?: string | null;
	/**
	 * The roles of the user.
	 */
	roles?: ('user' | 'editor' | 'superadmin')[] | null;
	persona?: (string | null) | Persona;
	sub?: string | null;
	stripeID?: string | null;
	skipSync?: boolean | null;
	updatedAt: string;
	createdAt: string;
	email: string;
	resetPasswordToken?: string | null;
	resetPasswordExpiration?: string | null;
	salt?: string | null;
	hash?: string | null;
	_verified?: boolean | null;
	_verificationToken?: string | null;
	loginAttempts?: number | null;
	lockUntil?: string | null;
	sessions?:
		| {
				id: string;
				createdAt?: string | null;
				expiresAt: string;
		  }[]
		| null;
	password?: string | null;
}
// ...
export interface Location {
	id: string;
	name: string;
	image?: (string | null) | Media;
	address?: string | null;
	mapUrl?: string | null;
	phone?: string | null;
	email?: string | null;
	website?: string | null;
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "activity".
 */
export interface Activity {
	id: string;
	name: string;
	shortDescription: string;
	description: {
		root: {
			type: string;
			children: {
				type: any;
				version: number;
				[k: string]: unknown;
			}[];
			direction: ('ltr' | 'rtl') | null;
			format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
			indent: number;
			version: number;
		};
		[k: string]: unknown;
	};
	relatedPersonas?:
		| {
				role?: ('docente' | 'speaker' | 'assistente' | 'tutor') | null;
				persona?: (string | null) | Persona;
				id?: string | null;
		  }[]
		| null;
	program?: {
		/**
		 * Where will the activity be held?
		 */
		where?: ('online' | 'irl') | null;
		onlineLink?: string | null;
		location?: (string | null) | Location;
		/**
		 * List of appointments
		 */
		appointments?:
			| {
					start: string;
					end: string;
					location?: (string | null) | Location;
					onlineLink?: string | null;
					shortDescription?: string | null;
					id?: string | null;
			  }[]
			| null;
	};
	enrollment?: {
		enrollable?: boolean | null;
		activityStatus?:
			| ('draft' | 'enrolling' | 'enrolled' | 'confirmed' | 'cancelled' | 'archived')
			| null;
		start?: string | null;
		deadline?: string | null;
		minParticipants?: number | null;
		maxParticipants?: number | null;
		requirePayment?: boolean | null;
		price?: number | null;
		/**
		 * Enable an introduction section to the form
		 */
		enableIntro?: boolean | null;
		introContent?: {
			root: {
				type: string;
				children: {
					type: any;
					version: number;
					[k: string]: unknown;
				}[];
				direction: ('ltr' | 'rtl') | null;
				format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
				indent: number;
				version: number;
			};
			[k: string]: unknown;
		} | null;
		form?: (string | null) | Form;
	};
	enrollments?: {
		docs?: (string | Enrollment)[];
		hasNextPage?: boolean;
		totalDocs?: number;
	};
	meta?: {
		title?: string | null;
		/**
		 * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
		 */
		image?: (string | null) | Media;
		description?: string | null;
	};
	slug?: string | null;
	slugLock?: boolean | null;
	image: string | Media;
	type?: ('event' | 'course' | 'workshop' | 'party' | 'open_call' | 'other') | null;
	topics?: (string | Topic)[] | null;
	video?: (string | null) | Video;
	year_month?: string[] | null;
	updatedAt: string;
	createdAt: string;
	_status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "form".
 */
export interface Form {
	id: string;
	name: string;
	description?: string | null;
	fields?:
		| (
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						defaultValue?: string | null;
						options?:
							| {
									label: string;
									value: string;
									description?: string | null;
									id?: string | null;
							  }[]
							| null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'select';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						required?: boolean | null;
						defaultValue?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'checkbox';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						defaultValue?: string | null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'text';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						defaultValue?: string | null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'textarea';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'email';
				  }
				| {
						message?: {
							root: {
								type: string;
								children: {
									type: any;
									version: number;
									[k: string]: unknown;
								}[];
								direction: ('ltr' | 'rtl') | null;
								format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
								indent: number;
								version: number;
							};
							[k: string]: unknown;
						} | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'message';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						defaultValue?: number | null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'number';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'country';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'state';
				  }
				| {
						name: string;
						label?: string | null;
						description?: string | null;
						width?: number | null;
						maxFiles?: number | null;
						required?: boolean | null;
						id?: string | null;
						blockName?: string | null;
						blockType: 'file';
				  }
		  )[]
		| null;
	submitButtonLabel?: string | null;
	/**
	 * Send custom emails when the form submits. Use comma separated lists to send the same email to multiple recipients. To reference a value from this form, wrap that field's name with double curly brackets, i.e. {{firstName}}.
	 */
	emails?:
		| {
				emailTo?: string | null;
				cc?: string | null;
				bcc?: string | null;
				replyTo?: string | null;
				emailFrom?: string | null;
				subject: string;
				/**
				 * Enter the message that should be sent in this email.
				 */
				message?: {
					root: {
						type: string;
						children: {
							type: any;
							version: number;
							[k: string]: unknown;
						}[];
						direction: ('ltr' | 'rtl') | null;
						format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
						indent: number;
						version: number;
					};
					[k: string]: unknown;
				} | null;
				id?: string | null;
		  }[]
		| null;
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "enrollment".
 */
export interface Enrollment {
	id: string;
	activity: string | Activity;
	user: string | User;
	formSubmission: string | FormSubmission;
	paymentId?: string | null;
	status?:
		| (
				| 'draft'
				| 'pending_payment'
				| 'payment_failed'
				| 'awaiting_admin_confirmation'
				| 'confirmed'
				| 'cancelled'
				| 'completed'
		  )
		| null;
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "form-submission".
 */
export interface FormSubmission {
	id: string;
	form: string | Form;
	submissionData?:
		| {
				field: string;
				value: string;
				id?: string | null;
		  }[]
		| null;
	user?: (string | null) | User;
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "video".
 */
export interface Video {
	id: string;
	title: string;
	description?: string | null;
	activities?: {
		docs?: (string | Activity)[];
		hasNextPage?: boolean;
		totalDocs?: number;
	};
	thumbnail?: (string | null) | Media;
	prefix?: string | null;
	updatedAt: string;
	createdAt: string;
	url?: string | null;
	thumbnailURL?: string | null;
	filename?: string | null;
	mimeType?: string | null;
	filesize?: number | null;
	width?: number | null;
	height?: number | null;
	focalX?: number | null;
	focalY?: number | null;
}